fixed all problems with unit and integration tests
